backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib[bcrypt]
python-jose
